Patch native compilation warnings.

+++ src/validation.cc
@@ -68,7 +68,7 @@ static int isLegalUTF8(const uint8_t *source, const int length)
 int is_valid_utf8 (size_t len, char *value)
 {
   /* is the string valid UTF-8? */
-  for (int i = 0; i < len; i++) {
+  for (size_t i = 0; i < len; i++) {
     uint32_t ch = 0;
     uint8_t  extrabytes = trailingBytesForUTF8[(uint8_t) value[i]];
